+++ x-pack/platform/plugins/shared/actions/docs/connector_spec/connector_spec_examples/abuseipdb.ts
@@ -32,7 +32,7 @@ export const AbuseIPDBConnector: SingleFileConnectorDefinition = {
     supportedFeatureIds: ['alerting', 'security'],
   },
 
-  authSchema: z.discriminatedUnion('method', [
+  schema: z.discriminatedUnion('method', [
     z.object({
       method: z.literal('headers'),
       headers: z.object({
