+++ x-pack/platform/plugins/shared/actions/docs/connector_spec/connector_spec_examples/alienvault_otx.ts
@@ -32,7 +32,7 @@ export const AlienVaultOTXConnector: SingleFileConnectorDefinition = {
     supportedFeatureIds: ['alerting', 'security'],
   },
 
-  authSchema: z.discriminatedUnion('method', [
+  schema: z.discriminatedUnion('method', [
     z.object({
       method: z.literal('headers'),
       headers: z.object({
