+++ x-pack/platform/plugins/shared/actions/docs/connector_spec/connector_spec_examples/index.ts
